+++ app/teams/services/team_membership_service.py
@@ -4,6 +4,9 @@ from rest_framework.exceptions import ValidationError, PermissionDenied
 
 from app.teams.models import TeamMembership
 
+from core.permissions.base import get_team_role
+from core.constants.team_constant import TEAM_ROLE_HIERARCHY
+
 logger = logging.getLogger(__name__)
 
 
@@ -22,11 +25,25 @@ def add_team_member(*, team, user, role):
     return membership
 
 
-def remove_team_member(*, team, user):
+def remove_team_member(*, team, user, performed_by):
+    import ipdb; ipdb.set_trace()
     logger.info(f"Removing member {user.email} from team: {team.name}")
     if user == team.created_by:
         logger.warning(f"Attempt to remove team creator from team: {team.name}")
         raise PermissionDenied("Cannot remove team creator")
+
+    target_user_role = get_team_role(user, team)
+    action_user_role = get_team_role(performed_by, team)
+    
+    if TEAM_ROLE_HIERARCHY[target_user_role] >= TEAM_ROLE_HIERARCHY[action_user_role]:
+        logger.warning(f"Attempt to remove user with equal or higher role from team: {team.name}")
+        raise PermissionDenied("Cannot remove user with equal or higher role")
+    
+    if target_user_role == "MANAGER":
+        manager_count = team.memberships.filter(role="MANAGER").count()
+        if manager_count == 1:
+            logger.warning(f"Last manager attempted removal from team: {team.name}")
+            raise PermissionDenied("Last manager cannot be removed")
 
     membership = TeamMembership.objects.get(team=team, user=user)
     membership.delete()
@@ -35,13 +52,21 @@ def remove_team_member(*, team, user):
 
 def self_remove_team_member(*, team, user):
     logger.info(f"Self-remove requested by {user.email} from team: {team.name}")
-    if not team.is_self_remove_allowed:
+    user_role = get_team_role(user, team)
+    
+    if not team.settings.allow_self_removal:
         logger.warning(f"Self-remove not allowed for team: {team.name}")
         raise PermissionDenied("Self removal not allowed")
 
     if user == team.created_by:
-        logger.warning(f"Team creator attempted self-remove from team: {team.name}")
-        raise PermissionDenied("Team creator cannot remove themselves")
+        logger.warning(f"Team owner attempted self-remove from team: {team.name}")
+        raise PermissionDenied("Team owner cannot remove themselves")
+    
+    if user_role == "MANAGER":
+        manager_count = team.membership.filter(role="MANAGER").count()
+        if manager_count == 1:
+            logger.warning(f"Last manager attempted self-remove from team: {team.name}")
+            raise PermissionDenied("Last manager cannot remove themselves.")
 
     membership = TeamMembership.objects.get(team=team, user=user)
     membership.delete()
